Add files via upload
This project intends to predict fraudulent credit card transactions with the help of machine learning models. The dataset is taken from the Kaggle website and it has a total of 2,84,807 transactions, out of which 492 are fraudulent. Since the dataset is highly imbalanced, so various balancing techniques have been applied to handle the imbalanced data. After finding the best evaluation score on test data, cost-benefit analysis has also been performed for the final predictive model.